version bump V7.0.1

+++ src/cryptonote_core/cryptonote_core.h
@@ -810,7 +810,7 @@ namespace cryptonote
       * requested range.  The optional value will be empty only if requesting the full chain *and*
       * another thread is already calculating it.
       */
-     std::optional<std::tuple<int64_t, int64_t, int64_t>> get_coinbase_tx_sum(uint64_t start_offset, size_t count);
+     std::optional<std::tuple<unsigned long long, unsigned long long, unsigned long long>> get_coinbase_tx_sum(uint64_t start_offset, size_t count);
 
      /**
       * @brief get the network type we're on
